+++ extension/Changelog.md
@@ -1,6 +1,15 @@
 # BCT Changelog
 
+## 0.6.2
+### Bugfixes:
+- Fixed a bug with BF Locks which made them unavailable to unlock in public rooms
+
+### Changed:
+- Added a reset button for all settings
+- Item permissions for BF Locks are now handled in BCTweaks settings page under Best Friends
+
 ## 0.6.1
+### Bugfixes:
 - Update for R101
 - Fixed an error with BF Lock where player wouldn't get access without Room Share enabled